==> simulations/notebooks_sim_bin/0.4_sim_ind_compLasso_binary_update.ipynb <==
### summarize compositional lasso results on Independent Simulation Scenarios for binary outcome

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

dim.list = list()
size = c(50, 100, 500, 1000)
idx = 0
for (P in size){
    for (N in size){
        idx = idx + 1
        dim.list[[idx]] = c(P=P, N=N)
    }
}

files = NULL
for (dim in dim.list){
    p = dim[1]
    n = dim[2]
    files = cbind(files, paste0(dir, '/sim_independent_', paste('P', p, 'N', n, sep='_'), '.RData'))
}

length(files)

avg_FDR = NULL
table_toe = NULL
tmp_num_select = rep(0, length(files))
for (i in 1:length(files)){
    print(paste0('indx: ', i))
    load(paste0(dir, '/binary_update/ind_GenCompLasso_binary_', i, '.RData')) 
    
    table_toe = rbind(table_toe, results_ind_GenCompLasso[c('n', 'p', 'rou', 'FP', 'FN', 'ROC', 'Stab')])
    tmp_num_select[i] = mean(rowSums(results_ind_GenCompLasso$Stab.table))
    
    # calculate FDR
    load(file_name, dat <- new.env())
    sub = dat$sim_array[[i]]
    p = sub$p # take true values from 1st replicate of each simulated data
    coef = sub$beta
    coef.true = which(coef != 0)
    
    tt = results_ind_GenCompLasso$Stab.table
    FDR = NULL # false positive rate
    for (r in 1:nrow(tt)){
        FDR = c(FDR, length(setdiff(which(tt[r, ] !=0), coef.true))/sum(tt[r, ]))

    }
    
    avg_FDR = c(avg_FDR, mean(FDR, na.rm=T))
}
table_toe = as.data.frame(table_toe)
table_toe$num_select = tmp_num_select
table_toe$FDR = round(avg_FDR,2)

head(table_toe)

tail(table_toe)

# export result
result.table_toe <- apply(table_toe,2,as.character)
rownames(result.table_toe) = rownames(table_toe)
result.table_toe = as.data.frame(result.table_toe)

# extract numbers only for 'n' & 'p'
result.table_toe$n = tidyr::extract_numeric(result.table_toe$n)
result.table_toe$p = tidyr::extract_numeric(result.table_toe$p)
result.table_toe$ratio = result.table_toe$p / result.table_toe$n

result.table_toe = result.table_toe[c('n', 'p', 'rou', 'ratio', 'Stab', 'ROC', 'FP', 'FN', 'num_select', 'FDR')]
colnames(result.table_toe)[1:4] = c('N', 'P', 'Corr', 'Ratio')

# convert interested measurements to be numeric
result.table_toe$Stab = as.numeric(as.character(result.table_toe$Stab))
# result.table_toe$ROC_mean = as.numeric(substr(result.table_toe$ROC, start=1, stop=4))
# result.table_toe$FP_mean = as.numeric(substr(result.table_toe$FP, start=1, stop=4))
# result.table_toe$FN_mean = as.numeric(substr(result.table_toe$FN, start=1, stop=4))
# result.table_toe$FN_mean[is.na(result.table_toe$FN_mean)] = 0
result.table_toe$num_select = as.numeric(as.character(result.table_toe$num_select))

result.table_toe$ROC_mean = as.numeric(sub("\\(.*", "", result.table_toe$ROC))
result.table_toe$FP_mean = as.numeric(sub("\\(.*", "", result.table_toe$FP))
result.table_toe$FN_mean = as.numeric(sub("\\(.*", "", result.table_toe$FN))

# check whether missing values exists
result.table_toe[rowSums(is.na(result.table_toe)) > 0,]

head(result.table_toe)

tail(result.table_toe)

result.table_toe

## export
write.table(result.table_toe, '../results_summary_updated/sim_ind_GencompLasso_binary.txt', sep='\t', row.names=F)
