==> simulations/notebooks_sim_bin/2.4_sim_block_GencompLasso_binary_update.ipynb <==
### summarize compositional lasso results on Block Simulation Scenarios for binary outcome

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data'

dim.list = list()
size = c(50, 100, 500, 1000)
idx = 0
for (P in size){
    for (N in size){
        idx = idx + 1
        dim.list[[idx]] = c(P=P, N=N)
    }
}

rou.list = seq(0.1, 0.9, 0.2)

files = NULL
for (rou in rou.list){
  for (dim in dim.list){
    p = dim[1]
    n = dim[2]
    files = cbind(files, paste0(dir, '/sim_block_corr', rou, paste('P', p, 'N', n, sep='_'), '.RData', sep=''))
  }
}

length(files)

avg_FDR = NULL
table_block = NULL
tmp_num_select = rep(0, length(files))
for (i in 1:length(files)){
    print(paste0('indx: ', i))
    load(paste0(dir, '/binary_update/block_GenCompLasso_binary_', i, '.RData')) 
    
    table_block = rbind(table_block, results_block_GenCompLasso[c('n', 'p', 'rou', 'FP', 'FN', 'ROC', 'Stab')])
    tmp_num_select[i] = mean(rowSums(results_block_GenCompLasso$Stab.table))
    
    # calculate FDR
    load(file_name, dat <- new.env())
    sub = dat$sim_array[[i]]
    p = sub$p # take true values from 1st replicate of each simulated data
    coef = sub$beta
    coef.true = which(coef != 0)
    
    tt = results_block_GenCompLasso$Stab.table
    FDR = NULL # false positive rate
    for (r in 1:nrow(tt)){
        FDR = c(FDR, length(setdiff(which(tt[r, ] !=0), coef.true))/sum(tt[r, ]))

    }
    
    avg_FDR = c(avg_FDR, mean(FDR, na.rm=T))
}
table_block = as.data.frame(table_block)
table_block$num_select = tmp_num_select
table_block$FDR = round(avg_FDR,2)

head(table_block)

tail(table_block)

# export result
result.table_block <- apply(table_block,2,as.character)
rownames(result.table_block) = rownames(table_block)
result.table_block = as.data.frame(result.table_block)

# extract numbers only for 'n' & 'p'
result.table_block$n = tidyr::extract_numeric(result.table_block$n)
result.table_block$p = tidyr::extract_numeric(result.table_block$p)
result.table_block$ratio = result.table_block$p / result.table_block$n

result.table_block = result.table_block[c('n', 'p', 'rou', 'ratio', 'Stab', 'ROC', 'FP', 'FN', 'num_select', 'FDR')]
colnames(result.table_block)[1:4] = c('N', 'P', 'Corr', 'Ratio')

# convert interested measurements to be numeric
result.table_block$Stab = as.numeric(as.character(result.table_block$Stab))
result.table_block$ROC_mean = as.numeric(sub("\\(.*", "", result.table_block$ROC))
result.table_block$FP_mean = as.numeric(sub("\\(.*", "", result.table_block$FP))
result.table_block$FN_mean = as.numeric(sub("\\(.*", "", result.table_block$FN))
result.table_block$num_select = as.numeric(as.character(result.table_block$num_select))

# check whether missing values exists
result.table_block[rowSums(is.na(result.table_block)) > 0,]

head(result.table_block)

tail(result.table_block)

result.table_block

## export
write.table(result.table_block, '../results_summary_bin/sim_block_GenCompLasso_binary.txt', sep='\t', row.names=F)
