==> simulations/notebooks_sim_bin/3_sim_binary_summary_update.ipynb <==
### summarize method comparison results across 4 feature selection methods in different simulation scenarios

### summarize Independent results

ind_lasso = read.csv('../results_summary_bin/sim_ind_lasso_binary.txt', sep='\t')
ind_elnet = read.csv('../results_summary_bin/sim_ind_Elnet_binary.txt', sep='\t')
ind_rf = read.csv('../results_summary_bin/sim_ind_RF_binary.txt', sep='\t')
ind_compLasso = read.csv('../results_summary_bin/sim_ind_GencompLasso_binary.txt', sep='\t')

ind_lasso$method = rep('lasso', dim(ind_lasso)[1])
ind_elnet$method = rep('elnet', dim(ind_elnet)[1])
ind_rf$method = rep('rf', dim(ind_rf)[1])
ind_compLasso$method = rep('compLasso', dim(ind_compLasso)[1])

dim(ind_lasso)
dim(ind_elnet)
dim(ind_rf)
dim(ind_compLasso)

ind = rbind(ind_lasso, ind_elnet)
ind = rbind(ind, ind_rf)
ind = rbind(ind, ind_compLasso)
dim(ind)

ind[ind$num_select == 0, ]

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
ind$FP_mean[ind$num_select ==0] = 0
write.table(ind, '../results_summary_bin/table_ind_binary_all.txt', sep='\t', row.names=F)

### summarize Toeplitz results

toe_lasso = read.csv('../results_summary_bin/sim_toe_lasso_binary.txt', sep='\t')
toe_elnet = read.csv('../results_summary_bin/sim_toe_Elnet_binary.txt', sep='\t')
toe_rf = read.csv('../results_summary_bin/sim_toe_RF_binary.txt', sep='\t')
toe_compLasso = read.csv('../results_summary_bin/sim_toe_GencompLasso_binary.txt', sep='\t')

toe_lasso$method = rep('lasso', dim(toe_lasso)[1])
toe_elnet$method = rep('elnet', dim(toe_elnet)[1])
toe_rf$method = rep('rf', dim(toe_rf)[1])
toe_compLasso$method = rep('compLasso', dim(toe_compLasso)[1])

dim(toe_lasso)
dim(toe_elnet)
dim(toe_rf)
dim(toe_compLasso)

toe = rbind(toe_lasso, toe_elnet)
toe = rbind(toe, toe_rf)
toe = rbind(toe, toe_compLasso)
dim(toe)

toe[toe$num_select == 0, ]

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
toe$FP_mean[toe$num_select ==0] = 0
write.table(toe, '../results_summary_bin/table_toe_binary_all.txt', sep='\t', row.names=F)

### summarize Block results

block_lasso = read.csv('../results_summary_bin/sim_block_lasso_binary.txt', sep='\t')
block_elnet = read.csv('../results_summary_bin/sim_block_Elnet_binary.txt', sep='\t')
block_rf = read.csv('../results_summary_bin/sim_block_RF_binary.txt', sep='\t')
block_compLasso = read.csv('../results_summary_bin/sim_block_GenCompLasso_binary.txt', sep='\t')

block_lasso$method = rep('lasso', dim(block_lasso)[1])
block_elnet$method = rep('elnet', dim(block_elnet)[1])
block_rf$method = rep('rf', dim(block_rf)[1])
block_compLasso$method = rep('compLasso', dim(block_compLasso)[1])

dim(block_lasso)
dim(block_elnet)
dim(block_rf)
dim(block_compLasso)

block = rbind(block_lasso, block_elnet)
block = rbind(block, block_rf)
block = rbind(block, block_compLasso)
dim(block)

block[block$num_select ==0, ] 

# fix false positives: when nothing was selected, then fp should be zero
# fp is now 1 due to length(setdiff(0/NA)) = 1 when actually should be 0 (in cv_sim_apply.R)
block$FP_mean[block$num_select ==0] = 0
write.table(block, '../results_summary_bin/table_block_binary_all.txt', sep='\t', row.names=F)
