==> simulations/notebooks_sim_bin/4_sim_binary_hypothesis_testing.ipynb <==
### Hypothesis testing based on boostrap for selected simulation scenario

dir = '/panfs/panfs1.ucsd.edu/panscratch/lij014/Stability_2020/sim_data/'

load(paste0(dir, 'binary_update/boot_toe_RF_binary.RData'))
load(paste0(dir, 'binary_update/boot_block_RF_binary.RData'))
load(paste0(dir, 'binary_update/boot_toe_genCompLasso_binary.RData'))
load(paste0(dir, 'binary_update/boot_block_genCompLasso_binary.RData'))

### hypothesis testing based on Stability

table = as.data.frame(cbind(toe_rf$stab_index, toe_genCompLasso$stab_index, block_rf$stab_index, block_genCompLasso$stab_index))
colnames(table) = c('toe_rf', 'toe_genCompLasso', 'block_rf', 'block_genCompLasso')
head(table)

mean(table$toe_genCompLasso)
mean(table$toe_rf)

diff_toe = (table$toe_genCompLasso - table$toe_rf)
mean(diff_toe)
quantile(diff_toe, probs = c(0.025, 0.975))

mean(table$block_genCompLasso)
mean(table$block_rf)

diff_block = (table$block_genCompLasso - table$block_rf)
mean(diff_block)
quantile(diff_block, probs = c(0.025, 0.975))

### hypothesis testing based on ROC

#### block 0.5, p = 1000, n = 100

i = 46
load(paste0(dir, '/binary_update/block_RF_binary_', i, '.RData')) 
results_block_rf[c('n', 'p', 'rou')]
block_rf_rocs = results_block_rf$ROC.list

i = 46
load(paste0(dir, '/binary_update/block_GenCompLasso_binary_', i, '.RData')) 
results_block_GenCompLasso[c('n', 'p', 'rou')]
block_GenCompLasso_rocs = results_block_GenCompLasso$ROC.list

diff_block = (block_GenCompLasso_rocs - block_rf_rocs)
mean(diff_block)
quantile(diff_block, probs = c(0.025, 0.975))

table(diff_block)

#### Toeplitz 0.5

i = 46
load(paste0(dir, '/binary_update/toe_RF_binary_', i, '.RData')) 
results_toe_rf[c('n', 'p', 'rou')]
toe_rf_rocs = results_toe_rf$ROC.list

i = 46
load(paste0(dir, '/binary_update/toe_GenCompLasso_binary_', i, '.RData')) 
results_toe_GenCompLasso[c('n', 'p', 'rou')]
toe_GenCompLasso_rocs = results_toe_GenCompLasso$ROC.list

diff_toe = (toe_GenCompLasso_rocs - toe_rf_rocs)
mean(diff_toe)
quantile(diff_toe, probs = c(0.025, 0.975))

table(diff_toe)
